api_requirements.txt
seaborn
fastapi
moviepy==1.0.3
torchaudio
pydub
noisereduce
uvicorn
textgrid
python-multipart==0.0.18
dragonmapper==0.3.0
hanziconv==0.3.2
spacy-pkuseg==1.0.0

zhconv
fasttext-wheel; sys_platform == 'darwin' or sys_platform == 'windows'